+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Changed
+- Consecutive streaming outputs on `stdout` in a cell output context (vice widget context) will be "squashed" into a single output before passed to `.add_cell_content`
 
 ## [0.5.5] - 2023-04-26
 ### Fixed